chore: Rename command from patch to export
