+++ redskyctl/internal/commands/commands.go
@@ -29,12 +29,12 @@ import (
 	"github.com/redskyops/redskyops-controller/redskyctl/internal/commands/configure"
 	"github.com/redskyops/redskyops-controller/redskyctl/internal/commands/docs"
 	"github.com/redskyops/redskyops-controller/redskyctl/internal/commands/experiments"
+	"github.com/redskyops/redskyops-controller/redskyctl/internal/commands/export"
 	"github.com/redskyops/redskyops-controller/redskyctl/internal/commands/generate"
 	"github.com/redskyops/redskyops-controller/redskyctl/internal/commands/grant_permissions"
 	"github.com/redskyops/redskyops-controller/redskyctl/internal/commands/initialize"
 	"github.com/redskyops/redskyops-controller/redskyctl/internal/commands/kustomize"
 	"github.com/redskyops/redskyops-controller/redskyctl/internal/commands/login"
-	"github.com/redskyops/redskyops-controller/redskyctl/internal/commands/patch"
 	"github.com/redskyops/redskyops-controller/redskyctl/internal/commands/ping"
 	"github.com/redskyops/redskyops-controller/redskyctl/internal/commands/reset"
 	"github.com/redskyops/redskyops-controller/redskyctl/internal/commands/results"
@@ -85,7 +85,7 @@ func NewRedskyctlCommand() *cobra.Command {
 	rootCmd.AddCommand(kustomize.NewCommand())
 	rootCmd.AddCommand(version.NewCommand(&version.Options{Config: cfg}))
 	rootCmd.AddCommand(docs.NewCommand(&docs.Options{}))
-	rootCmd.AddCommand(patch.NewCommand(&patch.Options{Config: cfg}))
+	rootCmd.AddCommand(export.NewCommand(&export.Options{Config: cfg}))
 
 	// TODO Add 'backup' and 'restore' maintenance commands ('maint' subcommands?)
 	// TODO We need helpers for doing a "dry run" on patches to make configuration easier
