+++ redskyctl/internal/commands/export/export_util_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package patch_test
+package export_test
 
 import (
 	"io/ioutil"
